go/proto/py_script_server/requirements.txt
grpcio==1.62.1
protobuf==5.26.1
